+++ CHANGELOG.md
@@ -14,6 +14,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 * Update to jsonwebtoken 3.
 * All frontend demos also send name to server on "Enter" keypress.
+* Demo backend CLI supports changing host and port.
+* Demo backend CLI uses default JWT secret when run on loopback.
 
 ### Fixed